requirements.txt
streamlit
tflite-runtime
Pillow
numpy
tensorflow-cpu==2.15.0
